# backend/openapi.yaml
openapi: 3.0.3
info:
  title: AiSaaS API
  version: 0.1.0
  description: |
    Route-level OpenAPI contract for the backend service.

    This file tracks registered paths, HTTP methods, authentication schemes,
    common response envelopes, and path/query parameters. Field-level request
    and response details remain documented in ../API接口文档.md.
servers:
  - url: http://127.0.0.1:8080
    description: Local development backend
...
paths:
...
  /api/admin/apps:
    get:
      tags: [admin]
      summary: List applications
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/IncludeHistoryQuery"
      responses:
        "200":
          $ref: "#/components/responses/ApplicationList"
        default:
          $ref: "#/components/responses/ApiError"
    post:
      tags: [admin]
      summary: Create application
      security:
        - AdminSession: []
          AdminCsrf: []
      requestBody:
        $ref: "#/components/requestBodies/ApplicationCreate"
      responses:
        "200":
          $ref: "#/components/responses/ApplicationCreate"
        default:
          $ref: "#/components/responses/ApiError"
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: admin_session
    AdminRefresh:
      type: apiKey
      in: cookie
      name: admin_refresh
    AdminCsrf:
      type: apiKey
      in: header
      name: X-CSRF-Token
    ClientBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceId:
      type: apiKey
      in: header
      name: X-Device-Id
    DeviceKeyId:
      type: apiKey
      in: header
      name: X-Device-Key-Id
    DeviceTimestamp:
      type: apiKey
      in: header
      name: X-Timestamp
    DeviceNonce:
      type: apiKey
      in: header
      name: X-Nonce
    DeviceBodySha256:
      type: apiKey
      in: header
      name: X-Body-SHA256
    DeviceSignature:
      type: apiKey
      in: header
      name: X-Signature
...
  parameters:
...
    IncludeHistoryQuery:
      name: include_history
      in: query
      required: false
      schema:
        type: boolean
        default: false
...
  requestBodies:
...
    ApplicationCreate:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApplicationCreateRequest"
...
  responses:
...
    ApplicationList:
      description: Wrapped application list response
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiEnvelope"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ApplicationListData"
...
    ApplicationCreate:
      description: Wrapped application creation response
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiEnvelope"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ApplicationCreateData"
...
    ApiError:
      description: Wrapped error API response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiEnvelope"
  schemas:
    ApiEnvelope:
      type: object
      required: [code, message, data, request_id]
      properties:
        code:
          type: integer
          format: int32
          example: 0
        errorCode:
          type: string
          nullable: true
          description: Stable string error code on error responses. Usually omitted on success.
          example: model_not_support_reference_video
        message:
          type: string
          example: ok
        data:
          nullable: true
          description: Endpoint-specific payload. See ../API接口文档.md.
        request_id:
          type: string
          example: req_01HX0000000000000000000000
...
    ApplicationSummary:
      type: object
      required:
        [
          id,
          name,
          slug,
          app_key,
          auth_mode,
          status,
          heartbeat_interval_seconds,
          offline_tolerance_seconds,
          max_devices_default,
          metadata,
        ]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
          nullable: true
        app_key:
          type: string
        auth_mode:
          type: string
          enum: [license, subscription, both]
        status:
          type: string
          enum: [active, disabled, archived]
        heartbeat_interval_seconds:
          type: integer
          format: int32
          minimum: 1
        offline_tolerance_seconds:
          type: integer
          format: int32
          minimum: 1
        max_devices_default:
          type: integer
          format: int32
          minimum: 0
        metadata:
          type: object
          additionalProperties: true
    SigningKey:
      type: object
      required:
        [
          id,
          kid,
          key_scope,
          alg,
          public_key_pem,
          status,
          not_before,
          not_after,
          created_at,
          activated_at,
        ]
      properties:
        id:
          type: string
          format: uuid
        kid:
          type: string
        key_scope:
          type: string
          enum: [app_request, release_file, secure_script, jwt_access_token]
        alg:
          type: string
          enum: [EdDSA]
        public_key_pem:
          type: string
        status:
          type: string
          enum: [active, retiring, retired, revoked]
        not_before:
          type: string
          format: date-time
        not_after:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        activated_at:
          type: string
          format: date-time
          nullable: true
    ApplicationListData:
      type: object
      required: [items, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApplicationSummary"
        meta:
          $ref: "#/components/schemas/ApplicationListMeta"
    ApplicationListMeta:
      type: object
      required: [page, page_size]
      properties:
        page:
          type: integer
          format: int32
          minimum: 1
        page_size:
          type: integer
          format: int32
          minimum: 1
          maximum: 100
...
    ApplicationCreateData:
      type: object
      required: [id, app_key, app_secret, signing_key, application]
      properties:
        id:
          type: string
          format: uuid
        app_key:
          type: string
        app_secret:
          type: string
        signing_key:
          $ref: "#/components/schemas/SigningKey"
        application:
          $ref: "#/components/schemas/ApplicationSummary"
...
    ApplicationCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
          nullable: true
        auth_mode:
          type: string
          nullable: true
          enum: [license, subscription, both]
        heartbeat_interval_seconds:
          type: integer
          format: int32
          minimum: 1
          nullable: true
        offline_tolerance_seconds:
          type: integer
          format: int32
          minimum: 1
          nullable: true
        max_devices_default:
          type: integer
          format: int32
          minimum: 0
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
...